requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
colorama==0.4.6
cryptography==42.0.5
Django==4.2.11
django-cloudinary-storage==0.3.0
django-elasticsearch-dsl==7.2.1
django-haystack==3.2.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.13.1
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.0.0
h11==0.14.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
PySocks==1.7.1
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tenacity==8.2.3
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wsproto==1.2.0
